refactor: rename topic helper file for V1
